Mayhem Heroes Initial Integration
