+++ mayhem/Mayhemfile
@@ -1,0 +1,8 @@
+project: defusedxml
+target: parse-fuzz
+testsuite:
+    - file://mayhem/testsuite
+
+cmds:
+  - cmd: /src/mayhem/fuzz_parse.py
+    libfuzzer: true
