--- mayhem/testsuite/0ab8318acaf6e678dd02e2b5c343ed41111b393d
@@ -1,1 +1,0 @@
-!
