--- mayhem/testsuite/0d1ae655b8a8eca94778537f43db08bac870abf5
@@ -1,1 +1,0 @@
-!--C
