--- mayhem/testsuite/203b0cb07d104f2244fe9015aa938e75824c4862
@@ -1,11 +1,0 @@
-<!-- Tested with xalan-j_2_7_1pizib-.n, Xerces-J-bin.2.11.0.tar.gz on
-     OpenJDK 1.7.0_15
-
-    $ LC_ALL=C java -cp xalan.jar:serializer.jar:xercesImpl.jar:xml-apis.jar \
-      org.apache.xalan.xslt.Process -in simple.xml -xsl xalan_write.xsl
--->
-<xsl:stylesheet version="1.0"
-    xmlns:xsl="h:write>
-  </xsl:template>
-</xsl:stylesheet>
-
