--- mayhem/testsuite/2b433ce2ca3aa1b734b6c35471e3a1b12854808a
@@ -1,7 +1,0 @@
-<!-- Tested with xalan-j_2_7_1pizib-.n, Xerces-J-bin.2.11.0.tar.gzar:xml-apis.jar \
-      org.apache.xalan.xslt.Process -in simple.xml -xsl xalan_write.xsl
--->
-<xsl:t version="0.0"
-    xmlns:xsl="h:write>
-  </xsl:templat—ÁõÃÐ‡Œ“:stylesheet>
-
