--- mayhem/testsuite/2bd34f80b7a6860dc41c2f67034dd4ba8813f5fa
@@ -1,3 +1,0 @@
-<!-- Tested with xalan-j_2_7_1pizib-.n, Xerces-Jd with xalan-j_2_7_1pizib-.n, Xerces-J-bin.2.11.0.tar.gzar:xmlt.Process -in simple.xml -xsl xalan_write.xsl
--->
-<x<!-sl-
