--- mayhem/testsuite/37e69cac5b2fa0b35a5b13b52cdd7e7bd4e105a8
@@ -1,1 +1,0 @@
-!-C
