--- mayhem/testsuite/3e065894094291ebadbbcf4a35c9b40282cf4a8d
@@ -1,8 +1,0 @@
-<!-- Tested with xalan-j_2_7_1pizib-.n, Xerces-J-bin.2.11.0.tar.gzar:xml-apis.jar \
-      org.apache.xalan.xslt.Process -in simple.xml -xsl xalan_write.xsl
--->
-<xsl:stylesheet version="0.0"
-    xmlns:xsl="h:write>
-  </xsl:template>
-</xsl:stylesheet>
-
