--- mayhem/testsuite/53164b4ce04a69739f15fedf4ea686742ef7ada9
@@ -1,1 +1,0 @@
-<!-re -TPo!eil xsl-
