--- mayhem/testsuite/540cc776a3c6a095e5f680a77f4debf582a6cbcf
@@ -1,1 +1,0 @@
-!C
