--- mayhem/testsuite/7664c205e90a53c8bec9c16c36fc558da03c7396
@@ -1,14 +1,0 @@
-<!-- Tested with xalan-j_2_7_1pizib-.n, Xerces-J-bin.2.11.0.tar.gz on
-     OpenJDK 1.7.0_15
-
-    $ LC_ALL=C java -cp xalan.jar:serializer.jar:xercesImpl.jar:xml-apis.jar \
-      org.apache.xalan.xslt.Process -in simple.xml -xsl xalan_write.xsl
--->
-<xsl:stylesheet version="1.0"
-    xmlns:xsl="http://www.w3.org/1999/XSL/Transform"
-    xmlns:redirect="http://xml.apache.org/xalan/redirect"
-    extension-element-prefixes="redtext>Something bad happens here!&#13;</xsl:text>
-    </redirect:write>
-  </xsl:template>
-</xsl:stylesheet>
-
