--- mayhem/testsuite/833122ab4864b68f5cb8089efa0b99285627fe26
@@ -1,1 +1,0 @@
-<!-- Tested wrar:xmlt.Process -il xsl-
